QueryWrapperUtil中对map2obj方法进行强化 	1.新增 在封装数据前，根据实体类属性类型对数据进行强制转换 	2.解决之前Long\String\Integer等不同类型之间无法转换的情况 Student业务逻辑修改： 	1.新增Student,则将student_id作为User的username和password,创建对应的User 	2.新增Student创建User时，为User绑定普通用户的权限 	3.删除Student时，对应解除上述新增的所有内容 User业务逻辑修改： 	1.新增User,需要指定并绑定权限 	2.删除User,需要解绑权限 待处理： 	1.Student删除时，根据Student表项主键id查询其他内容时的处理逻辑，有待优化 	2.权限管理存在问题，需要从头解决

+++ Graduate/src/main/java/com/txy/graduate/controller/MenuController.java
@@ -53,7 +53,7 @@ public class MenuController {
     }
 
     @DeleteMapping("{menu_id}")
-    public Result deleteSysMenuByMid(@PathVariable Integer menu_id){
+    public Result deleteSysMenuByMid(@PathVariable Long menu_id){
         boolean flag = menuService.deleteSysMenuById(menu_id);
         return Result.result(flag,null,flag?"删除成功 ^_^":"删除失败 -_-");
     }
@@ -72,9 +72,7 @@ public class MenuController {
 
         // 获取权限信息
         String authorityInfo = userService.getUserAuthorityInfo(sysUser.getId());
-        //if (authorityInfo.endsWith(",")){
-        //    authorityInfo = authorityInfo.substring(0,authorityInfo.length());
-        //}
+
         String[] authorityArr = authorityInfo.split(",");
 
         // 获取导航栏
